Replaced dft() free function with DFT struct
