+++ src/algorithm/dft.rs
@@ -57,8 +57,22 @@ impl<T: FFTnum> FFTAlgorithm<T> for DFT<T> {
 #[cfg(test)]
 mod unit_tests {
     use super::*;
+    use std::f32;
     use test_utils::{random_signal, compare_vectors};
-    use dft;
+    use num::{Complex, Zero};
+
+    fn dft(signal: &[Complex<f32>], spectrum: &mut [Complex<f32>]) {
+        for (k, spec_bin) in spectrum.iter_mut().enumerate() {
+            let mut sum = Zero::zero();
+            for (i, &x) in signal.iter().enumerate() {
+                let angle = -1f32 * (i * k) as f32 * 2f32 * f32::consts::PI / signal.len() as f32;
+                let twiddle = Complex::from_polar(&1f32, &angle);
+
+                sum = sum + twiddle * x;
+            }
+            *spec_bin = sum;
+        }
+    }
 
     #[test]
     fn test_matches_dft() {
@@ -68,8 +82,8 @@ mod unit_tests {
             dft(&input, &mut expected);
 
             let mut actual = input.clone();
-            let wrapper = DFT::new(len, false);
-            wrapper.process(&mut input, &mut actual);
+            let dft_instance = DFT::new(len, false);
+            dft_instance.process(&mut input, &mut actual);
 
             assert!(compare_vectors(&expected, &actual), "length = {}", len);
         }
@@ -80,5 +94,74 @@ mod unit_tests {
         let mut zero_output: Vec<Complex<f32>> = Vec::new();
 
         zero_dft.process(&mut zero_input, &mut zero_output);
+    }
+
+    /// Returns true if our `dft` function calculates the given spectrum from the
+    /// given signal, and if rustfft's DFT struct does the same
+    fn test_dft_correct(signal: &[Complex<f32>], spectrum: &[Complex<f32>]) -> bool {
+        assert_eq!(signal.len(), spectrum.len());
+
+        let expected_signal = signal.to_vec();
+        let mut expected_spectrum = vec![Zero::zero(); spectrum.len()];
+
+        let mut actual_signal = signal.to_vec();
+        let mut actual_spectrum = vec![Zero::zero(); spectrum.len()];
+
+        dft(&expected_signal, &mut expected_spectrum);
+
+        let dft_instance = DFT::new(signal.len(), false);
+        dft_instance.process(&mut actual_signal, &mut actual_spectrum);
+
+        return compare_vectors(spectrum, &expected_spectrum) && compare_vectors(spectrum, &actual_spectrum);
+    }
+
+    #[test]
+    fn test_dft_known_len_2() {
+        let signal = [Complex{re: 1f32, im: 0f32},
+                      Complex{re:-1f32, im: 0f32}];
+        let spectrum = [Complex{re: 0f32, im: 0f32},
+                        Complex{re: 2f32, im: 0f32}];
+        assert!(test_dft_correct(&signal[..], &spectrum[..]));
+    }
+
+    #[test]
+    fn test_dft_known_len_3() {
+        let signal = [Complex{re: 1f32, im: 1f32},
+                      Complex{re: 2f32, im:-3f32},
+                          Complex{re:-1f32, im: 4f32}];
+        let spectrum = [Complex{re: 2f32, im: 2f32},
+                        Complex{re: -5.562177f32, im: -2.098076f32},
+                        Complex{re: 6.562178f32, im: 3.09807f32}];
+        assert!(test_dft_correct(&signal[..], &spectrum[..]));
+    }
+
+    #[test]
+    fn test_dft_known_len_4() {
+        let signal = [Complex{re: 0f32, im: 1f32},
+                      Complex{re: 2.5f32, im:-3f32},
+                      Complex{re:-1f32, im: -1f32},
+                      Complex{re: 4f32, im: 0f32}];
+        let spectrum = [Complex{re: 5.5f32, im: -3f32},
+                        Complex{re: -2f32, im: 3.5f32},
+                        Complex{re: -7.5f32, im: 3f32},
+                        Complex{re: 4f32, im: 0.5f32}];
+        assert!(test_dft_correct(&signal[..], &spectrum[..]));
+    }
+
+    #[test]
+    fn test_dft_known_len_6() {
+        let signal = [Complex{re: 1f32, im: 1f32},
+                      Complex{re: 2f32, im: 2f32},
+                      Complex{re: 3f32, im: 3f32},
+                      Complex{re: 4f32, im: 4f32},
+                      Complex{re: 5f32, im: 5f32},
+                      Complex{re: 6f32, im: 6f32}];
+        let spectrum = [Complex{re: 21f32, im: 21f32},
+                        Complex{re: -8.16f32, im: 2.16f32},
+                        Complex{re: -4.76f32, im: -1.24f32},
+                        Complex{re: -3f32, im: -3f32},
+                        Complex{re: -1.24f32, im: -4.76f32},
+                        Complex{re: 2.16f32, im: -8.16f32}];
+        assert!(test_dft_correct(&signal[..], &spectrum[..]));
     }
 }
